Moved BiomeProviderRegisterEvent to event package, better ThreadLocal handling, event phase
